Refactor: merge models files into schema files, adjust unit test, change creating related products field to contain only products names (this will be extended with proper up-to-date data while retrieving specific product details from API); Change Prettier EOL to LF;
