+++ .prettierrc
@@ -4,7 +4,7 @@
   "tabWidth": 2,
   "semi": true,
   "printWidth": 120,
-  "endOfLine": "crlf",
+  "endOfLine": "lf",
   "overrides": [
     {
       "files": "*.js",
